App Service Manage task review comments (#6143)
* resolved review comments #1

* resolved review comments #2

* addressed review comments #3

* comment mock tests

* addressed review comments #4

* addressed review comments #5

* resolve build failures

* remove inherit

* remove unreachable codes & Fix few tests

* Fix L0 Tests

* addressed review comments

* changed case

* change source to target

* removed xml2js dependency
