+++ Tasks/AzureAppServiceManage/azureappservicemanage.ts
@@ -8,53 +8,10 @@ import { AzureApplicationInsights } from 'azure-arm-rest/azure-arm-appinsights';
 import { Kudu } from 'azure-arm-rest/azure-arm-app-service-kudu';
 import { ApplicationInsightsWebTests } from 'azure-arm-rest/azure-arm-appinsights-webtests';
 import { Resources } from 'azure-arm-rest/azure-arm-resource';
-
-const APPLICATION_INSIGHTS_EXTENSION_NAME: string = "Microsoft.ApplicationInsights.AzureWebSites";
-const pingApplicationCount: number = 1;
-const productionSlot: string = "production";
-
-async function enableContinuousMonitoring(appService: AzureAppService, appInsights: AzureApplicationInsights) {
-    var appDetails = await appService.get();
-    var appInsightsResource = await appInsights.get();
-    var appInsightsWebTests = new ApplicationInsightsWebTests(appInsights.getEndpoint(), appInsights.getResourceGroupName());
-    var webDeployPublishingProfile = await appService.getWebDeployPublishingProfile();
-    var applicationUrl = webDeployPublishingProfile.destinationAppUrl;
-    if(appDetails.kind.indexOf("linux") == -1) {
-        var appKuduService = await appService.getKuduService();
-        await appKuduService.installSiteExtension(APPLICATION_INSIGHTS_EXTENSION_NAME);
-    }
-
-    appInsightsResource.tags["hidden-link:" + appDetails.id] = "Resource";
-    tl.debug('Link app insights with app service via tag');
-    await appInsights.update(appInsightsResource);
-    tl.debug('Link app service with app insights via instrumentation key');
-    await appService.patchApplicationSettings({"APPINSIGHTS_INSTRUMENTATIONKEY": appInsightsResource.properties['InstrumentationKey']});
-    try {
-        tl.debug('Enable alwaysOn property for app service.');
-        await appService.patchConfiguration({"alwaysOn": true});    
-    }
-    catch(error) {
-        tl.warning(error);
-    }
-    
-    try {
-        tl.debug('add web test for app service - app insights');
-        await appInsightsWebTests.addWebTest(appInsightsResource, applicationUrl);
-    }
-    catch(error) {
-        tl.warning(error);
-    }
-}
-
-
-async function updateDeploymentStatusInKudu(kuduService: Kudu, taskResult: boolean, DeploymentID: string, customMessage: any) {
-    try {
-        return await kuduService.updateDeployment(taskResult, DeploymentID, customMessage);
-    }
-    catch(error) {
-        tl.warning(error);
-    }
-}
+import { AzureAppServiceUtils } from './operations/AzureAppServiceUtils';
+import { KuduServiceUtils } from './operations/KuduServiceUtils';
+import { AzureResourceFilterUtils } from './operations/AzureResourceFilterUtils';
+import { enableContinuousMonitoring } from './operations/ContinuousMonitoringUtils';
 
 async function run() {
     try {
@@ -76,91 +33,78 @@ async function run() {
         var taskResult = true;
         var errorMessage: string = "";
         var updateDeploymentStatus: boolean = true;
-
         var azureEndpoint: AzureEndpoint = await new AzureRMEndpoint(connectedServiceName).getEndpoint();
-        var resources: Array<any> = await new Resources(azureEndpoint).getResources('Microsoft.Web/Sites', webAppName);
 
         if(action != "Swap Slots" && !slotName) {
-            if(!resources || resources.length == 0) {
-                throw new Error(tl.loc('ResourceDoesntExist', webAppName));
-            }
-            else if(resources.length == 1) {
-                resourceGroupName = resources[0].id.split("/")[4];
-            }
-            else {
-                throw new Error(tl.loc('MultipleResourceGroupFoundForAppService', webAppName));
-            }
+            resourceGroupName = await AzureResourceFilterUtils.getResourceGroupName(azureEndpoint, 'Microsoft.Web/Sites', webAppName);
         }
 
-        tl.debug(`Resource Group: ${resourceGroupName}`);    
+        tl.debug(`Resource Group: ${resourceGroupName}`);
+        var appService: AzureAppService = new AzureAppService(azureEndpoint, resourceGroupName, webAppName, slotName);
+        var azureAppServiceUtils: AzureAppServiceUtils = new AzureAppServiceUtils(appService);
+
         switch(action) {
             case "Start Azure App Service": {
-                var appService: AzureAppService = new AzureAppService(azureEndpoint, resourceGroupName, webAppName, slotName);
                 await appService.start();
-                await appService.monitorAppState("running");
-                await appService.pingApplication(pingApplicationCount);
+                await azureAppServiceUtils.monitorApplicationState("running");
+                await azureAppServiceUtils.pingApplication();
                 break;
             }
             case "Stop Azure App Service": {
-                var appService: AzureAppService = new AzureAppService(azureEndpoint, resourceGroupName, webAppName, slotName);
                 await appService.stop();
-                await appService.monitorAppState("stopped");
-                await appService.pingApplication(pingApplicationCount);
+                await azureAppServiceUtils.monitorApplicationState("stopped");
                 break;
             }
             case "Restart Azure App Service": {
-                var appService: AzureAppService = new AzureAppService(azureEndpoint, resourceGroupName, webAppName, slotName);
                 await appService.restart();
+                await azureAppServiceUtils.pingApplication();
                 break;
             }
             case "Swap Slots": {
-                targetSlot = (swapWithProduction) ? productionSlot : targetSlot;
+                targetSlot = (swapWithProduction) ? "production" : targetSlot;
                 var appServiceSourceSlot: AzureAppService = new AzureAppService(azureEndpoint, resourceGroupName, webAppName, sourceSlot);
                 var appServiceTargetSlot: AzureAppService = new AzureAppService(azureEndpoint, resourceGroupName, webAppName, targetSlot);
+                var appServiceSourceSlotUtils: AzureAppServiceUtils = new AzureAppServiceUtils(appServiceSourceSlot);
+                var appServiceTargetSlotUtils: AzureAppServiceUtils = new AzureAppServiceUtils(appServiceTargetSlot);
+
                 if(appServiceSourceSlot.getSlot().toLowerCase() == appServiceTargetSlot.getSlot().toLowerCase()) {
                     updateDeploymentStatus = false;
                     throw new Error(tl.loc('SourceAndTargetSlotCannotBeSame'));
-
                 }
+
                 console.log(tl.loc('WarmingUpSlots'));
-                await appServiceSourceSlot.pingApplication(1);
-                await appServiceTargetSlot.pingApplication(1);
+                try {
+                    await Promise.all([appServiceSourceSlotUtils.pingApplication(), appServiceTargetSlotUtils.pingApplication()]);
+                }
+                catch(error) {
+                    tl.debug('Failed to warm-up slots. Error: ' + error);
+                }
+
                 await appServiceSourceSlot.swap(targetSlot, preserveVnet);
                 break;
             }
             case "Start all continuous webjobs": {
-                var appService: AzureAppService = new AzureAppService(azureEndpoint, resourceGroupName, webAppName, slotName);
-                var appServiceKuduService = await appService.getKuduService();
-                console.log(tl.loc('StartingContinousWebJobs'));
-                await appServiceKuduService.startContinuousWebJobs();
-                console.log(tl.loc('StartedContinousWebJobs'));
+                var appServiceKuduService: Kudu = await azureAppServiceUtils.getKuduService();
+                var kuduServiceUtils: KuduServiceUtils = new KuduServiceUtils(appServiceKuduService);
+                await kuduServiceUtils.startContinuousWebJobs();
                 break;
             }
             case "Stop all continuous webjobs": {
-                var appService: AzureAppService = new AzureAppService(azureEndpoint, resourceGroupName, webAppName, slotName);
-                var appServiceKuduService = await appService.getKuduService();
-                console.log(tl.loc('StoppingContinousWebJobs'));
-                await appServiceKuduService.stopContinuousWebJobs();
-                console.log(tl.loc('StoppedContinousWebJobs'));
+                var appServiceKuduService = await azureAppServiceUtils.getKuduService();
+                var kuduServiceUtils: KuduServiceUtils = new KuduServiceUtils(appServiceKuduService);
+                await kuduServiceUtils.stopContinuousWebJobs();
                 break;
             }
             case "Install Extensions": {
-                var appService: AzureAppService = new AzureAppService(azureEndpoint, resourceGroupName, webAppName, slotName);
-                var appServiceKuduService = await appService.getKuduService();
+                var appServiceKuduService = await azureAppServiceUtils.getKuduService();
+                var kuduServiceUtils: KuduServiceUtils = new KuduServiceUtils(appServiceKuduService);
                 var extensionOutputVariablesArray = (extensionOutputVariables) ? extensionOutputVariables.split(',') : [];
-                await appServiceKuduService.installSiteExtensions(extensionList.split(','), extensionOutputVariablesArray);
+                await kuduServiceUtils.installSiteExtensions(extensionList.split(','), extensionOutputVariablesArray);
                 break;
             }
             case "Enable Continuous Monitoring": {
-                var appService: AzureAppService = new AzureAppService(azureEndpoint, resourceGroupName, webAppName, slotName);
                 var appInsights: AzureApplicationInsights = new AzureApplicationInsights(azureEndpoint, appInsightsResourceGroupName, appInsightsResourceName);
-                try {
-                    await enableContinuousMonitoring(appService, appInsights);
-                }
-                catch(error) {
-                    throw new Error(tl.loc('FailedToEnableContinuousMonitoring', error));
-                }
-                console.log(tl.loc("ContinousMonitoringEnabled", webAppName));
+                await enableContinuousMonitoring(azureEndpoint, appService, appInsights);
                 break;
             }
             default: {
@@ -177,22 +121,24 @@ async function run() {
     try {
         switch(action) {
             case "Swap Slots": {
-                if(appServiceSourceSlot && appServiceTargetSlot && updateDeploymentStatus) {
-                    var sourceSlotKuduService = await appServiceSourceSlot.getKuduService();
-                    var targetSlotKuduService = await appServiceTargetSlot.getKuduService();
+                if(appServiceSourceSlotUtils && appServiceTargetSlotUtils && updateDeploymentStatus) {
+                    var sourceSlotKuduService = await appServiceSourceSlotUtils.getKuduService();
+                    var targetSlotKuduService = await appServiceTargetSlotUtils.getKuduService();
+                    var sourceSlotKuduServiceUtils = new KuduServiceUtils(sourceSlotKuduService);
+                    var targetSlotKuduServiceUtils = new KuduServiceUtils(targetSlotKuduService);
                     var customMessage = {
                         'type': 'SlotSwap',
                         'sourceSlot': appServiceSourceSlot.getSlot(),
                         'targetSlot': appServiceTargetSlot.getSlot()
                     }
-                    var DeploymentID = await updateDeploymentStatusInKudu(sourceSlotKuduService, taskResult, null, customMessage);
-                    await updateDeploymentStatusInKudu(targetSlotKuduService, taskResult, DeploymentID, customMessage);
+                    var DeploymentID = await sourceSlotKuduServiceUtils.updateDeploymentStatus(taskResult, null, customMessage);
+                    await targetSlotKuduServiceUtils.updateDeploymentStatus(taskResult, DeploymentID, customMessage);
                 }
                 break;
             }
             case "Install Extensions": {
-                if(appServiceKuduService) {
-                    await updateDeploymentStatusInKudu(appServiceKuduService, taskResult, null, {"type": action});
+                if(kuduServiceUtils) {
+                    await kuduServiceUtils.updateDeploymentStatus(taskResult, null, { "type" : action });
                 }
                 break;
             }
