+++ Tasks/Common/azure-arm-rest/Tests/mock_utils.ts
@@ -1,6 +1,7 @@
 import { AzureEndpoint, WebTest } from '../azureModels';
 import { ApplicationInsightsWebTests } from '../azure-arm-appinsights-webtests';
 import * as querystring from "querystring";
+import { ApplicationTokenCredentials } from '../azure-arm-common';
 export var nock = require('nock');
 
 export function getMockEndpoint() {
@@ -15,7 +16,9 @@ export function getMockEndpoint() {
         tenantID: "MOCK_TENANT_ID",
         url: "https://management.azure.com/",
         environmentAuthorityUrl: "https://login.windows.net/",
-        activeDirectoryResourceID: "https://management.azure.com/"
+        activeDirectoryResourceID: "https://management.azure.com/",
+        applicationTokenCredentials: new ApplicationTokenCredentials("MOCK_SPN_ID", "MOCK_TENANT_ID", "MOCK_SPN_KEY", "https://management.azure.com/",
+        "https://login.windows.net/", "https://management.azure.com/", false)
     }
     
     nock("https://login.windows.net", {
@@ -37,9 +40,23 @@ export function getMockEndpoint() {
 }
 
 export function mockAzureARMAppInsightsWebTests() {
-    var MockWebTest1: WebTest = (new ApplicationInsightsWebTests(getMockEndpoint(), "MOCK_RESOURCE_GROUP_NAME")).configureNewWebTest("MOCK_APP_INSIGHTS_1", "http://MOCK_APP_1.azurewebsites.net", "MOCK_TEST_1");
+    var MockWebTest1 = {
+        type: '',
+        location: '',
+        tags: {},
+        name: 'MOCK_TEST_1',
+        id: "hidden-link:/subscriptions/MOCK_SUBSCRIPTION_ID/resourceGroups/MOCK_RESOURCE_GROUP_NAME/providers/microsoft.insights/components/MOCK_APP_INSIGHTS_1".toLowerCase()
+    };
+    
+    var MockWebTest2 = {
+        type: '',
+        location: '',
+        tags: {},
+        name: 'MOCK_TEST_2',
+        id: "hidden-link:/subscriptions/MOCK_SUBSCRIPTION_ID/resourceGroups/MOCK_RESOURCE_GROUP_NAME/providers/microsoft.insights/components/MOCK_APP_INSIGHTS_1".toLowerCase()
+    };
+
     MockWebTest1.tags = {"hidden-link:/subscriptions/MOCK_SUBSCRIPTION_ID/resourceGroups/MOCK_RESOURCE_GROUP_NAME/providers/microsoft.insights/components/MOCK_APP_INSIGHTS_1": "Resource"};
-    var MockWebTest2: WebTest = (new ApplicationInsightsWebTests(getMockEndpoint(), "MOCK_RESOURCE_GROUP_NAME")).configureNewWebTest("MOCK_APP_INSIGHTS_2", "http://MOCK_APP_2.azurewebsites.net", "MOCK_TEST_2");
     MockWebTest2.tags = {"hidden-link:/subscriptions/MOCK_SUBSCRIPTION_ID/resourceGroups/MOCK_RESOURCE_GROUP_NAME/providers/microsoft.insights/components/MOCK_APP_INSIGHTS_1": "Resource"};
 
     nock('https://management.azure.com', {
@@ -59,7 +76,6 @@ export function mockAzureARMAppInsightsWebTests() {
         "content-type": "application/json; charset=utf-8"
     }).put("/subscriptions/MOCK_SUBSCRIPTION_ID/resourceGroups/MOCK_RESOURCE_GROUP_NAME/providers/microsoft.insights/webtests/VSTS_MOCK_TEST_FAIL?api-version=2015-05-01")
     .reply(501, 'Failed to add new web test').persist();
-
 }
 
 export function mockAzureApplicationInsightsTests() {
