+++ Tasks/Common/azure-arm-rest/azure-arm-app-service.ts
@@ -3,7 +3,7 @@ import tl = require('vsts-task-lib/task');
 import util = require('util');
 import webClient = require('./webClient');
 import Q = require('q');
-import { 
+import {
     AzureEndpoint,
     AzureAppServiceConfigurationDetails
 } from './azureModels';
@@ -13,11 +13,10 @@ import {
     ToError
 } from './AzureServiceClient';
 import { Kudu } from './azure-arm-app-service-kudu';
-
+import constants = require('./constants');
 var parseString = require('xml2js').parseString;
 
 export class AzureAppService {
-    private _endpoint: AzureEndpoint;
     private _resourceGroup: string;
     private _name: string;
     private _slot: string;
@@ -27,17 +26,14 @@ export class AzureAppService {
     private _appServicePublishingProfile: any;
 
     constructor(endpoint: AzureEndpoint, resourceGroup: string, name: string, slot?: string, appKind?: string) {
-        var credentials = new msRestAzure.ApplicationTokenCredentials(endpoint.servicePrincipalClientID, endpoint.tenantID, endpoint.servicePrincipalKey, 
-            endpoint.url, endpoint.environmentAuthorityUrl, endpoint.activeDirectoryResourceID, endpoint.environment.toLowerCase() == 'azurestack');
-        this._client = new ServiceClient(credentials, endpoint.subscriptionID, 30);
-        this._endpoint = endpoint;
+        this._client = new ServiceClient(endpoint.applicationTokenCredentials, endpoint.subscriptionID, 30);
         this._resourceGroup = resourceGroup;
         this._name = name;
-        this._slot = (slot && slot.toLowerCase() == 'production') ? null : slot;
+        this._slot = (slot && slot.toLowerCase() == constants.productionSlot) ? null : slot;
         this._appKind = appKind;
     }
 
-    public async start() {
+    public async start(): Promise<void> {
         try {
             var webRequest = new webClient.WebRequest();
             webRequest.method = 'POST';
@@ -54,14 +50,13 @@ export class AzureAppService {
             }
 
             console.log(tl.loc('StartedAppService', this._getFormattedName()));
-            return response.statusCode;
         }
         catch(error) {
             throw Error(tl.loc('FailedToStartAppService', this._getFormattedName(), this._client.getFormattedError(error)));
         }
     }
 
-    public async stop() {
+    public async stop(): Promise<void> {
         try {
             var webRequest = new webClient.WebRequest();
             webRequest.method = 'POST';
@@ -78,14 +73,13 @@ export class AzureAppService {
             }
 
             console.log(tl.loc('StoppedAppService', this._getFormattedName()));
-            return response.statusCode;
         }
         catch(error) {
             throw Error(tl.loc('FailedToStopAppService', this._getFormattedName(), this._client.getFormattedError(error)));
         }
     }
 
-    public async restart() {
+    public async restart(): Promise<void> {
         try {
             var webRequest = new webClient.WebRequest();
             webRequest.method = 'POST';
@@ -102,14 +96,13 @@ export class AzureAppService {
             }
 
             console.log(tl.loc('RestartedAppService', this._getFormattedName()));
-            return response.body;
         }
         catch(error) {
             throw Error(tl.loc('FailedToRestartAppService', this._getFormattedName(), this._client.getFormattedError(error)));
         }
     }
 
-    public async swap(slotName: string, preserveVNet?: boolean) {
+    public async swap(slotName: string, preserveVNet?: boolean): Promise<void> {
         try {
             var webRequest = new webClient.WebRequest();
             webRequest.method = 'POST';
@@ -136,7 +129,6 @@ export class AzureAppService {
             }
 
             console.log(tl.loc('SwappedAppServiceSlotSlots', this._name, this.getSlot(), slotName));
-            return response.statusCode;
         }
         catch(error) {
             throw Error(tl.loc('FailedToSwapAppServiceSlotSlots', this._name, this.getSlot(), slotName, this._client.getFormattedError(error)));
@@ -151,87 +143,12 @@ export class AzureAppService {
         return this._appServiceConfigurationDetails;
     }
 
-    public async monitorAppState(state: string) {
-        state = state.toLowerCase();
-        if(["running", "stopped"].indexOf(state) == -1) {
-            throw new Error(tl.loc('InvalidMonitorAppState', state));
-        }
-
-        while(true) {
-            var appDetails = await this.get(true);
-            if(appDetails && appDetails.properties && appDetails.properties["state"]) {
-                tl.debug(`App Service state: ${appDetails.properties["state"]}`)
-                if(appDetails.properties["state"].toLowerCase() == state) {
-                    tl.debug(`App Service state '${appDetails.properties["state"]}' matched with expected state '${state}'.`);
-                    console.log(tl.loc('AppServiceState', appDetails.properties["state"]));
-                    break;
-                }
-                await webClient.sleepFor(5);
-            }
-            else {
-                tl.debug('Unable to monitor app service details as the state is unknown.');
-                break;
-            }
-        }
-    }
-
     public async getPublishingProfileWithSecrets(force?: boolean) {
         if(force || !this._appServicePublishingProfile) {
             this._appServicePublishingProfile = await this._getPublishingProfileWithSecrets();
         }
 
         return this._appServicePublishingProfile;
-    }
-
-    public async getWebDeployPublishingProfile() {
-        var publishingProfile = await this.getPublishingProfileWithSecrets();
-        var defer = Q.defer<any>();
-        parseString(publishingProfile, (error, result) => {
-            for (var index in result.publishData.publishProfile) {
-                if (result.publishData.publishProfile[index].$.publishMethod === "MSDeploy") {
-                    defer.resolve(result.publishData.publishProfile[index].$);
-                }
-            }
-            defer.reject(tl.loc('ErrorNoSuchDeployingMethodExists'));
-        });
-
-        return defer.promise;
-    }
-
-    public async pingApplication(numberOfTimes: number) {
-        numberOfTimes = numberOfTimes ? numberOfTimes : 1;
-        try {
-            var applicationUrl: string = (await this.getWebDeployPublishingProfile()).destinationAppUrl;    
-        }
-        catch(error) {
-            tl.debug(`Unable to get publishing profile for ping application. Error: ${this._client.getFormattedError(error)}`);
-        }
-        
-        if(!applicationUrl) {
-            tl.debug('Application Url not found.');
-            return;
-        }
-
-        tl.debug(`Ping App Service for '${numberOfTimes}' time(s).`);
-        var webRequest = new webClient.WebRequest();
-        webRequest.method = 'GET';
-        webRequest.uri = applicationUrl;
-
-        while(numberOfTimes > 0) {
-            try {
-                tl.debug('pausing for 5 seconds before request');
-                await webClient.sleepFor(5);
-                var response = await webClient.sendRequest(webRequest);
-
-                tl.debug(`App Service status Code: '${response.statusCode}'. Status Message: '${response.statusMessage}'`);
-            }
-            catch(error) {
-                tl.debug(`Unable to ping App Service. Error: ${this._client.getFormattedError(error)}`);
-            }
-            finally {
-                numberOfTimes -= 1;
-            }
-        }
     }
 
     public async getPublishingCredentials() {
@@ -257,7 +174,7 @@ export class AzureAppService {
         }
     }
 
-    public async getApplicationSettings() {
+    public async getApplicationSettings(): Promise<AzureAppServiceConfigurationDetails> {
         try {
             var httpRequest = new webClient.WebRequest();
             httpRequest.method = 'POST';
@@ -280,7 +197,7 @@ export class AzureAppService {
         }
     }
 
-    public async updateApplicationSettings(applicationSettings) {
+    public async updateApplicationSettings(applicationSettings): Promise<AzureAppServiceConfigurationDetails> {
         try {
             var httpRequest = new webClient.WebRequest();
             httpRequest.method = 'PUT';
@@ -304,7 +221,7 @@ export class AzureAppService {
         }
     }
 
-    public async patchApplicationSettings(properties) {
+    public async patchApplicationSettings(properties): Promise<void> {
         var applicationSettings = await this.getApplicationSettings();
         for(var key in properties) {
             applicationSettings.properties[key] = properties[key];
@@ -314,7 +231,7 @@ export class AzureAppService {
 
     }
     
-    public async getConfiguration() {
+    public async getConfiguration(): Promise<AzureAppServiceConfigurationDetails> {
         try {
             var httpRequest = new webClient.WebRequest();
             httpRequest.method = 'GET';
@@ -337,7 +254,7 @@ export class AzureAppService {
         }
     }
 
-    public async updateConfiguration(applicationSettings) {
+    public async updateConfiguration(applicationSettings): Promise<AzureAppServiceConfigurationDetails> {
         try {
             var httpRequest = new webClient.WebRequest();
             httpRequest.method = 'PUT';
@@ -361,7 +278,7 @@ export class AzureAppService {
         }
     }
 
-    public async patchConfiguration(properties) {
+    public async patchConfiguration(properties): Promise<void> {
         var applicationSettings = await this.getConfiguration();
         for(var key in properties) {
             applicationSettings.properties[key] = properties[key];
@@ -373,15 +290,6 @@ export class AzureAppService {
 
     public getSlot(): string {
         return this._slot ? this._slot : "production";
-    }
-
-    public async getKuduService() {
-        var publishingCredentials = await this.getPublishingCredentials();
-        if(publishingCredentials.properties["scmUri"]) {
-            tl.setVariable(`AZURE_APP_SERVICE_KUDU_${this._name}_${this.getSlot()}_PASSWORD`, publishingCredentials.properties["publishingPassword"], true);
-            return new Kudu(publishingCredentials.properties["scmUri"], publishingCredentials.properties["publishingUserName"], publishingCredentials.properties["publishingPassword"]);
-        }
-        throw Error(tl.loc('KuduSCMDetailsAreEmpty'));
     }
     
     private async _getPublishingProfileWithSecrets() {
@@ -435,4 +343,8 @@ export class AzureAppService {
     private _getFormattedName(): string {
         return this._slot ? `${this._name}-${this._slot}` : this._name;
     }
-}
+
+    public getName(): string {
+        return this._name;
+    }
+ }
