+++ Tasks/Common/azure-arm-rest/constants.ts
@@ -1,0 +1,7 @@
+export const AzureEnvironments = {
+    AzureStack: 'azurestack'
+};
+
+export const APPLICATION_INSIGHTS_EXTENSION_NAME: string = "Microsoft.ApplicationInsights.AzureWebSites";
+
+export const productionSlot: string = "production";
