+++ Tasks/Common/azure-arm-rest/package.json
@@ -15,7 +15,6 @@
   "dependencies": {
     "q": "1.4.1",
     "vsts-task-lib": "2.0.5",
-    "typed-rest-client": "0.12.0",
-    "xml2js": "0.4.13"
+    "typed-rest-client": "0.12.0"
   }
 }
